backend/requirements.txt
Flask
Flask-JWT-Extended
Werkzeug
pydub
soundfile
librosa
scikit-learn
torch<2.9
torchaudio<2.9
onnxruntime
transformers
accelerate
jiwer
pydantic
certifi
python-dotenv
flask-pymongo
flask-bcrypt
